requirements.txt
charset-normalizer==2.1.1
clip
coloredlogs==15.0.1
colorlog==6.9.0
cvxopt==1.3.0
Cython==0.29.32
editdistance==0.6.2
facenet-pytorch==2.5.2
fonttools==4.37.1
future==0.18.2
ftfy==6.1.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
huggingface-hub==0.16.4
imageio==2.21.2
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.12.0
info-nce-pytorch==0.1.4
insightface==0.7.3
ipython==7.34.0
joblib==1.1.0
Markdown==3.4.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.2
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.0
onnx==1.12.0
onnxruntime==1.10.0
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.34
packaging==21.3
pickleshare==0.7.5
Pillow
platformdirs==4.0.0
protobuf
pydub==0.25.1
Pygments==2.13.0
pyparsing==3.0.9
pytorch-lightning==1.7.5
pytorch-wpe==0.0.1
PyYAML==6.0
safetensors==0.4.3
scipy==1.7.3
six
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
timm==0.6.11
tokenizers==0.13.3
toml
torch-complex==0.4.4
torchmetrics==0.9.3
tqdm==4.64.1
transformers==4.30.2
typing_extensions==4.7.1
umap-learn==0.5.7
